+++ src/cljs/wombats_web_client/constants/local_storage.cljs
@@ -1,4 +1,8 @@
 (ns wombats-web-client.constants.local-storage)
 
+;; Users unique authentication token
 (def token "token")
 (def access-token "access-token")
+
+;; Used for new user enrollment
+(def access-key "access-key")